# docs/swagger.yaml
basePath: http://127.0.0.1:17187/
definitions:
  common.BaseInfo:
    properties:
      build_time:
        description: 构建时间
        type: string
      version:
        description: 版本号
        type: string
    type: object
  common.InfoResponse:
    properties:
      base:
        $ref: '#/definitions/common.BaseInfo'
      browser:
        $ref: '#/definitions/common.ToolInfo'
      git:
        $ref: '#/definitions/common.ToolInfo'
      location:
        type: string
      node:
        $ref: '#/definitions/common.ToolInfo'
      nvm:
        $ref: '#/definitions/common.ToolInfo'
      start_at:
        type: string
    type: object
  common.ToolInfo:
    properties:
      installed:
        type: boolean
      version:
        type: string
    type: object
  models.BotPackagesGitBranchCommitsInfo:
    properties:
      author:
        type: string
      date:
        type: string
      hash:
        type: string
      message:
        type: string
    type: object
  models.BotPackagesGitBranches:
    properties:
      branches:
        items:
          type: string
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_page:
        type: integer
    type: object
  models.BotPackagesGitCommits:
    properties:
      commits:
        items:
          $ref: '#/definitions/models.BotPackagesGitBranchCommitsInfo'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_page:
        type: integer
    type: object
  response.ResCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 1011
    - 1012
    - 1013
    - 1014
    - 1015
    - 1016
    - 1017
    - 1018
    - 1019
    format: int64
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeServerBusy
    - CodeInvalidParam
    - RobotCreateFailed
    - RobotNameIsEmpty
    - RobotNotExist
    - RobotAlreadyExist
    - RobotConfigNotExist
    - ReadRobotConfigFailed
    - ErrorRobotLog
    - CreateConfigFailed
    - ConfigNameIsEmpty
    - ConfigContentIsEmpty
    - ReadConfigFailed
    - ConfigFileIsDeleted
    - DeleteConfigFileFailed
    - ErrorLogLevel
    - ErrCreateConfigCatFailed
    - ErrReadConfigCatFailed
    - TokenInvalid
  response.ResponseData:
    properties:
      code:
        $ref: '#/definitions/response.ResCode'
      data: {}
      msg: {}
      request_code:
        type: integer
    type: object
host: 127.0.0.1:17187
info:
  contact: {}
  description: alemongo 机器人面板
  termsOfService: http://swagger.io/terms/
  title: 这里写标题
  version: 0.0.1
paths:
  /api/v1/bot/packages/git/branches:
    get:
      consumes:
      - application/json
      description: 获取应用git所有分支
      parameters:
      - description: bot名称
        in: query
        name: name
        required: true
        type: string
      - description: 应用名称
        in: query
        name: app_name
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.BotPackagesGitBranches'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 获取应用git所有分支
      tags:
      - 应用
  /api/v1/bot/packages/git/commits:
    get:
      consumes:
      - application/json
      description: 获取应用git分支提交记录
      parameters:
      - description: bot名称
        in: query
        name: name
        required: true
        type: string
      - description: 应用名称
        in: query
        name: app_name
        required: true
        type: string
      - description: 分支名称
        in: query
        name: branch_name
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.BotPackagesGitCommits'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 获取应用git分支提交记录
      tags:
      - 应用
  /api/v1/packages/switch:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 切换机器人应用到指定的 Git 分支和提交
      parameters:
      - description: 机器人名称
        in: formData
        name: name
        required: true
        type: string
      - description: 应用名称
        in: formData
        name: app_name
        required: true
        type: string
      - description: 分支名称
        in: formData
        name: branch_name
        required: true
        type: string
      - description: 提交哈希值
        in: formData
        name: commit_hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 切换成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                msg:
                  type: string
              type: object
        "400":
          description: 参数错误或应用不存在
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 切换分支失败
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                msg:
                  type: string
              type: object
      summary: 切换机器人应用分支
      tags:
      - 机器人应用
  /common/info:
    get:
      consumes:
      - application/json
      description: 获取环境信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.InfoResponse'
      summary: 获取环境信息
      tags:
      - 公共
  /user/login:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "429":
          description: 账户已被锁定
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 登陆接口
swagger: "2.0"